test: split actions test into specific files
